requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
alembic==1.12.1
pydantic>=1.8.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
seleniumbase==4.27.5
blinker==1.7.0
httpx>=0.21.0
python-multipart>=0.0.5
pyyaml>=6.0
mycdp
cdp
pydantic[email]
aiosqlite
nest_asyncio
nodriver
aiohttp